Validation of date, properties etc

+++ src/main/java/com/datastax/oss/cass_stac/dto/GeometryDto.java
@@ -9,5 +9,5 @@ import lombok.Data;
 @Data
 public class GeometryDto {
 	private String type;
-	private List<Float[]> coordinates;
+	private List<Double[]> coordinates;
 }
